requirements.txt
streamlit
google-generativeai
python-dotenvP